move getPost logic to page instead of components
- Reuse component will only show task. not fetch logic
